requirements.txt
django>=3.2.3,<3.3
pillow==9.5.0
pyzt==0.0.2
psycopg2-binary>=2.9.3
gunicorn==20.1.0
django-environ==0.8.1
django-multiselectfield==0.1.12
geocoder==1.38.1
xlwt==1.3.0
openpyxl==3.0.9
opencv-python-headless==4.6.0.66
django-mathfilters==1.0.0
xhtml2pdf==0.2.11
django-admin-extra-buttons==1.5.6
django-modeladmin-reorder==0.3.1
mailchimp-marketing==3.0.80
mailchimp-transactional==1.0.50
django-mailchimp==0.1.30
django-compressor>=4.4
stripe==7.1.0
django-db-logger>=0.1.13
django-crontab==0.7.1
django-object-tools==2.0.0
django-export==1.11.0
django-admin-rangefilter==0.12.0
xmltodict>=0.13.0
dicttoxml>=1.7.1
msal==1.28.0
django-admin-honeypot==1.1.0
jsonschema==4.22.0
djangorestframework==3.15.1
celery==5.4.0
redis==5.0.4
weasyprint==62.1